scripts/apply/pyproject.toml
[project]
name = "gaj-apply"
version = "0.1.0"
description = "GAJ ATS autofill subpackage. Python owns the browser; TypeScript owns the DB."
requires-python = ">=3.11"
dependencies = [
    "scrapling[fetchers]>=0.4.7",
    "PyYAML>=6.0",
    "beautifulsoup4>=4.14",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apply*"]
